% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{add_missing_column}
\alias{add_missing_column}
\title{add back samples with no CNA events}
\usage{
add_missing_column(name_order, output_data, val)
}
\description{
add back samples with no CNA events
}
