% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-score.R
\name{calculateTreeScoreMutations}
\alias{calculateTreeScoreMutations}
\title{Calculate SCHISM fitness scores for trees}
\usage{
calculateTreeScoreMutations(
  mcf_chain,
  data,
  icnTable,
  cncfTable,
  multiplicityTable,
  clusterAssingmentTable,
  purity,
  all_spanning_trees,
  restriction.val = 1
)
}
\description{
Calculate SCHISM fitness scores for trees
}
