% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-cna.R
\name{reassignCNA}
\alias{reassignCNA}
\title{assign integer copy number using updated CNCF and input tcn
max_cn: maximum integer copy number allowed}
\usage{
reassignCNA(cncf_update, tcn, max_cn = 8)
}
\description{
assign integer copy number using updated CNCF and input tcn
max_cn: maximum integer copy number allowed
}
