% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-cna.R
\name{reassignCNCF}
\alias{reassignCNCF}
\title{assign CNCF to the closest cluster using euclidean distance}
\usage{
reassignCNCF(cncf_est, w_chain)
}
\description{
assign CNCF to the closest cluster using euclidean distance
}
