% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{runGSEA}
\alias{runGSEA}
\title{GSEA analysis using fgsea}
\usage{
runGSEA(
  X_optimal,
  outputDir,
  treeFile,
  GSEA_file = NULL,
  top_K = 5,
  n_permutations = 1000
)
}
\description{
GSEA analysis using fgsea
}
