% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC-main.R
\name{runPictograph}
\alias{runPictograph}
\title{main function to infer tumor evolution history}
\usage{
runPictograph(
  mutation_file,
  copy_number_file = NULL,
  SNV_file = NULL,
  outputDir = NULL,
  sample_presence = TRUE,
  score = "silhouette",
  max_K = 10,
  min_mutation_per_cluster = 5,
  min_cluster_thresh = 0.05,
  cluster_diff_thresh = 0.05,
  n.iter = 5000,
  n.burn = 1000,
  thin = 10,
  mc.cores = 8,
  inits = list(.RNG.name = "base::Wichmann-Hill", .RNG.seed = 123),
  LOH = FALSE,
  purity_min = 0.2,
  driverFile = NULL,
  cytobandFile = NULL,
  alt_reads_thresh = 0,
  vaf_thresh = 0,
  cnv_min_length = 1e+06,
  tcn_normal_range = c(1.75, 2.3),
  filter_cnv = T,
  smooth_cnv = T,
  autosome = T,
  dual_model = TRUE,
  ploidy = 2,
  pval = 0.05,
  threshes = NULL
)
}
\arguments{
\item{mutation_file}{a csv file that include information for SSMs. See vignette for details.}

\item{copy_number_file}{a csv file that include information for CNA. See vignette for details.}

\item{SNV_file}{a csv file that include information for germline heterozygous SNVs. See vignette for details.}

\item{outputDir}{output directory for saving all files.}

\item{sample_presence}{whether or not to use sample presence to separate the mutations; default: TRUE}

\item{score}{scoring function to estimate the number of clusters. silhouette or BIC; default: silhuette}

\item{max_K}{user defined maximum number of clusters; default: 10}

\item{min_mutation_per_cluster}{minumum number of mutations in each cluster; default: 5}

\item{min_cluster_thresh}{minimum MCF for each cluster; default: 0.05}

\item{cluster_diff_thresh}{difference threshold to merge two clusters: default: 0.05}

\item{n.iter}{number of iterations by JAGS; default: 5000}

\item{n.burn}{number of burns by JAGS; default: 1000}

\item{thin}{number of thin by JAGS; default: 10}

\item{mc.cores}{number of cores to use for parallel computing; not applicable to windows; default: 8}

\item{inits}{additional parameters by JAGS.}

\item{LOH}{whether or not to include copy number segments that are copy neutral but LOH; default: FALSE}

\item{purity_min}{minimum purity for tumor samples; default: 0.2}

\item{driverFile}{list of driver genes used for visualization. See vignette for details.}

\item{cytobandFile}{list of cytoband regions used for visualization. See vignette for details.}

\item{alt_reads_thresh}{minimum number of alternative read count for a SSM to be included in the analysis; default: 0}

\item{vaf_thresh}{minimum VAF for a SSM to be included in the analysis; default: 0}

\item{cnv_min_length}{minimum length of copy number alterations for it to be included in analysis}

\item{tcn_normal_range}{range of total copy number considered as copy-neutral; default: c(1.75,2.3)}

\item{filter_cnv}{whether or not to filter copy number alterations; default: TRUE}

\item{smooth_cnv}{whether or not to process copy number alterations across samples to unify the segment start and end postions; default: TRUE}

\item{autosome}{to only include autosomes; default: TRUE}
}
\description{
run MCMC chains to  infer the clonal evolution of tumors from single or multi-region sequencing data. 
This function automatically runs a pipeline of the tool. It models uncertainty of mutation cellular 
fraction (MCF) in small somatic mutations (SSMs) and copy number alterations (CNAs), assigning SSMs 
and CNAs to subclones using a Bayesian hierarchical model, and reconstruct tumor evolutionary trees 
that are constrained based on principles of lineage precedence, sum condition, and optionally by 
sample-presence.
}
