% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-presence.R
\name{separateBySamplePresence}
\alias{separateBySamplePresence}
\title{separate mutation and copy number data based on sample presence profile}
\usage{
separateBySamplePresence(input_data)
}
\description{
separate mutation and copy number data based on sample presence profile
}
