% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-presence.R
\name{separateCNAsBySamplePresence}
\alias{separateCNAsBySamplePresence}
\title{separate copy number alterations by sample presence}
\usage{
separateCNAsBySamplePresence(input_data)
}
\description{
separate copy number alterations by sample presence
}
