% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-presence.R
\name{separateMutationsBySamplePresence}
\alias{separateMutationsBySamplePresence}
\title{separate mutations by sample presence}
\usage{
separateMutationsBySamplePresence(input_data)
}
\description{
separate mutations by sample presence
}
