% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{smoothCNV}
\alias{smoothCNV}
\title{merge segments with similar coordinates
the maximum allowed distance between the start or end position of two segments is the max(cnv_max_dist, min(length of two segments)*cnv_max_percent)}
\usage{
smoothCNV(data, cnv_max_dist = 2000, cnv_max_percent = 0.3)
}
\description{
merge segments with similar coordinates
the maximum allowed distance between the start or end position of two segments is the max(cnv_max_dist, min(length of two segments)*cnv_max_percent)
}
