% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{importCopyNumberFile}
\alias{importCopyNumberFile}
\title{import copy number file}
\usage{
importCopyNumberFile(
  copy_number_file,
  outputDir,
  SNV_file = NULL,
  LOH = FALSE,
  name_order = NULL,
  cnv_min_length = 1e+06,
  tcn_normal_range = c(1.75, 2.3),
  filter_cnv = T,
  smooth_cnv = F,
  autosome = T,
  pval = 0.05
)
}
\description{
import copy number file
}
