% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{importFiles}
\alias{importFiles}
\title{read input data file and store in PICTOGRAPH input format}
\usage{
importFiles(
  mutation_file,
  copy_number_file = NULL,
  alt_reads_thresh = 0,
  vaf_thresh = 0
)
}
\arguments{
\item{mutation_file}{mutation data file that contains columns "sample", "mutation", "chrom", "start", "end", total_reads", and "alt_reads";}

\item{copy_number_file}{copy number file that contains columns "sample", "chrom", "start", "end", "major", "minor"}
}
\description{
read input data file and store in PICTOGRAPH input format
}
