% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{importFiles}
\alias{importFiles}
\title{read input data file and store in PICTOGRAPH input format}
\usage{
importFiles(
  mutation_file,
  copy_number_file = NULL,
  SNV_file = NULL,
  outputDir = NULL,
  stat_file = NULL,
  alt_reads_thresh = 0,
  vaf_thresh = 0,
  cnv_max_dist = 2000,
  cnv_max_percent = 0.3,
  tcn_normal_range = c(1.8, 2.2),
  smooth_cnv = F,
  autosome = T,
  mc.cores = 8,
  pval = 0.05
)
}
\arguments{
\item{mutation_file}{mutation data file that contains columns "sample", "mutation", "chrom", "start", "end", total_reads", and "alt_reads";}

\item{copy_number_file}{copy number file that contains columns "sample", "chrom", "start", "end", "tcn"}
}
\description{
read input data file and store in PICTOGRAPH input format
}
