% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{importMutationFile}
\alias{importMutationFile}
\title{import mutation file}
\usage{
importMutationFile(
  mutation_file,
  alt_reads_thresh = 0,
  vaf_thresh = 0,
  purity_min = 0.2
)
}
\description{
import mutation file
}
