% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC-main.R
\name{mcmcMain}
\alias{mcmcMain}
\title{run MCMC with added subclonal copy number}
\usage{
mcmcMain(
  mutation_file,
  copy_number_file = NULL,
  SNV_file = NULL,
  outputDir = NULL,
  stat_file = NULL,
  sample_presence = FALSE,
  dual_model = TRUE,
  ploidy = 2,
  pval = 0.05,
  max_K = 5,
  min_mutation_per_cluster = 5,
  n.iter = 5000,
  n.burn = 1000,
  thin = 10,
  mc.cores = 8,
  inits = list(.RNG.name = "base::Wichmann-Hill", .RNG.seed = 123),
  cluster_diff_thresh = 0.05,
  alt_reads_thresh = 0,
  vaf_thresh = 0,
  cnv_max_dist = 2000,
  cnv_max_percent = 0.3,
  tcn_normal_range = c(1.8, 2.2),
  smooth_cnv = F,
  autosome = T
)
}
\description{
run MCMC with added subclonal copy number
}
