% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{runDeconvolution}
\alias{runDeconvolution}
\title{bulk RNA deconvolution using tumor evolution information}
\usage{
runDeconvolution(
  rna_file,
  treeFile,
  proportionFile,
  purityFile,
  outputDir,
  lambda = 0.2
)
}
\description{
bulk RNA deconvolution using tumor evolution information
}
